# refine use cache for mpt model