@yafshar 2024-07-25
@Jing1Ling can you check the other PR https://github.com/huggingface/optimum-habana/pull/1126, the code base has changed there