@Jing1Ling 2024-07-26
> @Jing1Ling can you check the other PR #1126, the code base has changed there

@yafshar Thank you for the reminder. I believe this patch is a more concise approach. It only changes a few lines of code, achieves lower memory usage and similar throughput.