@Jing1Ling 2024-07-26
**Why this change causes such a large decrease in memory usage**:
The conclusion is that 'query_states', 'key_states', and 'value_states' are all views of 'mixed_qkv'. When I create a new tensor tuple and copy 'key_states' and 'value_states' into it, 'mixed_qkv' will be released after the attention function returns. 

Before the modification, 'key_states' and 'value_states' were saved in 'past_key_value' as views of 'mixed_qkv', so 'mixed_qkv' could not be freed. In other words, with the change in this PR, each transformer layer (block) can save the space of one 'query_states' during the decode first token process(create k, v but release mixed_qkv, thus saving the space of one q).

I used torch.hpu.synchronize() and torch.hpu.memory_summary() to do memory statistics at each stage of decoding the first token. 
```bash
# exit after decoding the first token
python run_generation.py --model_name_or_path mosaicml/mpt-7b --use_hpu_graphs --use_kv_cache --limit_hpu_graph --batch_size 8 --max_input_tokens 128 --max_new_tokens 5 --trim_logits --attn_softmax_bf16 --warmup 3 --n_iterations 1 --bf16
```
Here is some evidence:

![image](https://github.com/user-attachments/assets/5c1e1ce8-b534-47c8-9a26-5b4ed3688e7f)
![image](https://github.com/user-attachments/assets/d135687a-64ad-4d99-84f3-3eca034e3727)