@yafshar 2024-07-26
@Jing1Ling thanks. 
@atakaha is this PR cover everything you added in https://github.com/huggingface/optimum-habana/pull/1126