@atakaha 2024-07-26
> @Jing1Ling thanks. @atakaha is this PR covers everything you added in #1126

In the point of performance, YES. In the point of "--reuse_cache" implementation like other modes, no.