@yafshar 2024-07-26
> > @Jing1Ling thanks. @atakaha is this PR covers everything you added in #1126
> 
> In the point of performance, YES. In the point of "--reuse_cache" implementation like other modes, no.

@atakaha can you tell me what is missing for `--reuse_cache` implementation? With the current implementation the memory usage seems the same and I think that is the main point of reuse_cache. So if you add that extra code what is the benefit? Am I missing anything?