@yafshar 2024-07-26
> Do you think any other model might benefit from this? As far as I can tell, mixed qkv is a feature of mpt/dbrx, and of the 2 we only have MPT here right?

Here is only MPT. I think the same can be useful for any similar model