@atakaha 2024-07-29
> > > > > @Jing1Ling thanks. @atakaha is this PR covers everything you added in #1126
> > > > 
> > > > 
> > > > In the point of performance, YES. In the point of "--reuse_cache" implementation like other modes, no.
> > > 
> > > 
> > > @atakaha can you tell me what is missing for `--reuse_cache` implementation? With the current implementation the memory usage seems the same and I think that is the main point of reuse_cache. So if you add that extra code what is the benefit? Am I missing anything?
> > 
> > 
> > The `--reuse_cache` implementation is followed other models, such llama, mistral approach. This PR introduce same performance of improvement of `--reuse_cache` implementation.
> 
> @atakaha thanks. Then please close #1126, let us merge this PR first and if there is any missing functionality we can start a new PR on top of this. Thanks for your contribution

@yafshar, https://github.com/huggingface/optimum-habana/pull/1126 is closed.